Various bug fixes (#31)
* Fixed integer overflow in allocation

* Fixed a number of different bugs in the GPU code, particularly sync of buffers

* Missing size_t

* Fixed issues

* Missing int

* Geometry init adjustments to work for ALL_COMMS

+++ Include/Geometry/geometry_omp.h
@@ -33,7 +33,7 @@
 #define _SITE_FOR_RED(type, ip, is, redop1, redop2) \
     _OMP_PRAGMA(_omp_parallel)                      \
     _OMP_PRAGMA(_omp_for redop1 redop2)             \
-    for (int is = (type)->master_start[ip]; is <= (type)->master_end[ip]; is++)
+    for (size_t is = (type)->master_start[ip]; is <= (type)->master_end[ip]; is++)
 
 /**
  * @brief Iterate over sites of a given piece.
